+++ test/types/schema.test.ts
@@ -305,7 +305,7 @@ function gh11439() {
   const bookSchema = new Schema<Book>({
     collection: String
   }, {
-    supressReservedKeysWarning: true
+    suppressReservedKeysWarning: true
   });
 }